yoyo_attacker/requirements.txt
requests==2.18.4
google-cloud==0.30.0
locustio==0.9.0
